Backend/restaurant_review/requirements.txt
asgiref==3.8.1
Django==5.1.2
django-cors-headers==4.6.0
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.8
inflection==0.5.1
packaging==24.2
psycopg2-binary==2.9.9
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
rest-framework-simplejwt==0.0.2
sqlparse==0.5.1
uritemplate==4.1.1
